+++ index.js
@@ -1,23 +1,8 @@
 'use strict'
 
-const fastJsonStringify = require('fast-json-stringify')
-
-function SerializerSelector () {
-  return function buildSerializerFactory (externalSchemas, serializerOpts) {
-    const fjsOpts = Object.assign({}, serializerOpts, { schema: externalSchemas })
-    return responseSchemaCompiler.bind(null, fjsOpts)
-  }
-}
-
-function responseSchemaCompiler (fjsOpts, { schema /* method, url, httpStatus */ }) {
-  if (fjsOpts.schema && schema.$id && fjsOpts.schema[schema.$id]) {
-    fjsOpts.schema = { ...fjsOpts.schema }
-    delete fjsOpts.schema[schema.$id]
-  }
-  return fastJsonStringify(schema, fjsOpts)
-}
+const { SerializerSelector, StandaloneSerializer } = require('./standalone')
 
 module.exports = SerializerSelector
 module.exports.default = SerializerSelector
 module.exports.SerializerSelector = SerializerSelector
-module.exports.StandaloneSerializer = require('./standalone')
+module.exports.StandaloneSerializer = StandaloneSerializer
